refactor: dissolve numpyro_forecast.functional into models.py, predictive.py, _offload.py, _validation.py
PR-C step C1 of the #75 refactor (refactor.md in #81): a pure layout move with no
behavior change. functional/models.py becomes numpyro_forecast/models.py;
functional/posterior.py and functional/prediction.py merge into predictive.py (which
retires the cross-module reach for the private _index_tree); _offload.py and
_validation.py move to the package root. The jaxtyping hook list, every import and
Sphinx role, the great-docs reference entries (models.* / predictive.*, sections
"Model building blocks" and "Producing draws"), and the test modules
(test_models.py, test_draw_posterior.py, test_predictive.py, git mv only) follow.

_WARNING_SKIP_PREFIXES took dirname(dirname(__file__)) because _offload.py lived one
level below the package; at the package root that would have skipped the whole
checkout (tests included) and misattributed the "cpu" fallback warning. It now uses
the file's own directory, and the attribution test asserts the prefix is exactly the
package directory. The device placement contract is documented once, on
draw_posterior; forecast, predict_in_sample and to_datatree point at it.

Also folded in from the plan's step 0: pre-commit hook revs aligned with the dev pins
(ruff v0.16.3, ty v0.0.72), `make store-durations` passes --clean-durations so
renamed test files drop their stale node ids, and .superpowers/ is git-ignored.
