+++ numpyro_forecast/__init__.py
@@ -13,9 +13,10 @@ with install_import_hook("numpyro_forecast", "beartype.beartype"):
         evaluate,
         exceptions,
         features,
-        functional,
         metrics,
+        models,
         optional,
+        predictive,
         surgery,
     )
 
@@ -42,17 +43,15 @@ from numpyro_forecast.exceptions import (
     VectorizedGuideError,
     VectorizedMetricError,
 )
-from numpyro_forecast.functional import (
+from numpyro_forecast.models import (
     Horizon,
     Transition,
-    draw_posterior,
-    forecast,
     markov_time_series,
     predict,
     predict_glm,
-    predict_in_sample,
     time_series,
 )
+from numpyro_forecast.predictive import draw_posterior, forecast, predict_in_sample
 from numpyro_forecast.surgery import register_elementwise
 
 try:
