+++ numpyro_forecast/typing.py
@@ -42,7 +42,7 @@ class ForecastModel(Protocol):
 
     Any plain function with this signature satisfies this Protocol structurally
     (for example, one that derives its
-    :class:`~numpyro_forecast.functional.models.Horizon` from the shapes via
+    :class:`~numpyro_forecast.models.Horizon` from the shapes via
     ``Horizon.from_data`` and calls the functional primitives), so nothing
     needs to subclass it. The parameters are positional-only so a user model's
     own parameter names (``cov``, ``y``, ...) stay free instead of being forced
